Thick Lines and Canvas
Although a lot of time was spent creating a nice, emissive texture for Thuraya, the line renderers were not happy with Fade. Therefore, the lines for the constellation have instead been made white and THICK. Lastly, the fading of Thuraya's canvas has been done, but it doesn't really fade, it pops away; needs fixing.